backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# RAG and LLM
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.16

# Vector Store
faiss-cpu==1.8.0

# Embeddings and LLM
ollama==0.1.6
httpx==0.26.0

# Document Processing
GitPython==3.1.41
chardet==5.2.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Optional: Hugging Face integration
huggingface-hub==0.20.3
